turn a console project
the project became a console project instead class library
