+++ cs/CircularCloudLayouter/TagsCloudVisualization1.csproj
@@ -43,7 +43,7 @@
   <ItemGroup>
     <Compile Include="CircularCloudLayouter.cs" />
     <Compile Include="CircularCloudLayouterTests.cs" />
-    <Compile Include="VisualisationTest.cs" />
+    <Compile Include="Programm.cs" />
     <Compile Include="Visualization.cs" />
   </ItemGroup>
   <ItemGroup>
