+++ cs/ConsoleApp2/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace ConsoleApp2
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
